pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Po Stevanus Andrianta <sandrianta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.82.0"
sendgrid = "^6.9.7"
twilio = "^7.14.0"
uvicorn = "^0.18.3"
firebase-admin = "^5.3.0"
ShopifyAPI = "^12.0.0"
google-cloud-secret-manager = "^2.12.5"
black = "^22.8.0"
fireo = "^1.6.1"
google-auth-oauthlib = "^0.7.1"
pytest = "^7.2.1"
httpx = "^0.23.3"
Jinja2 = "^3.1.2"
python-multipart = "^0.0.5"
google-cloud-storage = "^2.7.0"
currency-symbols = "^2.0.3"
pydantic = "^1.10.5"
pre-commit = "^3.1.1"
sentry-sdk = "^1.16.0"
python-dotenv = "^1.0.0"
boltons = "^23.0.0"
pyjwt = "^2.6.0"
aenum = "^3.1.11"
validators = "^0.20.0"
apscheduler = "^3.10.1"
fastapi-utils = "^0.2.1"
openai = "^0.27.7"
graphql-query = "^1.1.1"
dependency-injector = "^4.41.0"
pillow = "^9.5.0"
pyhumps = "^3.8.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:app"
